move auth function to auth0 file
