+++ client/src/auth0.ts
@@ -1,4 +1,5 @@
 import createAuth0Client, { Auth0Client } from '@auth0/auth0-spa-js';
+import auth from '@feathersjs/authentication-client';
 import determineConfigBasedOnEnvironment from './config';
 
 const config = determineConfigBasedOnEnvironment();
@@ -12,4 +13,31 @@ export async function setupAuth0(): Promise<Auth0Client> {
         audience: config.audience
     });
     return auth0;
+}
+
+export async function authenticate(app, client, callback): Promise<void> {
+    const auth0 = (await client.get('auth0')) as Auth0Client;
+    let redirectUrl = '';
+    const query = window.location.search;
+    if (query.includes('code=') && query.includes('state=')) {
+        const options = await auth0.handleRedirectCallback();
+        redirectUrl = options.appState.targetUrl;
+    }
+    if (await auth0.isAuthenticated()) {
+        client.configure(auth({ storage: window.sessionStorage }));
+        const accessToken = await auth0.getTokenSilently();
+        client.authentication.setAccessToken(accessToken);
+        const authResult = await client.reAuthenticate(true);
+        callback(authResult);
+
+        if (redirectUrl) {
+            const router = app.config.globalProperties.$router;
+            router.push(redirectUrl);
+        }
+    } else {
+        const options: RedirectLoginOptions = {
+            appState: { targetUrl: location.pathname + location.search },
+        };
+        await auth0.loginWithRedirect(options);
+    }
 }
